fix: numeric cast in enrollMemberInCycle + guard register welcome email
- enrollMemberInCycle: COUNT() is bigint, so $2 * COUNT coerced the
  monthly amount '500.00' to bigint and rejected the decimals, 500ing
  invitation acceptance (and rolling back the join). Cast to ::numeric.
- register: email.sendWelcome ran after res.json uncaught; if it throws
  it caused ERR_HTTP_HEADERS_SENT. Deferred + caught so post-response
  emails can never affect the sent response.

+++ src/services/chilimba.service.js
@@ -119,10 +119,12 @@ const enrollMemberInCycle = async (exec, groupId, userId) => {
        ON CONFLICT (group_id, cycle_number, payout_order) DO NOTHING`,
       [groupId, userId, cycle, order, payoutDate, g.monthly_amount]
     );
-    // Refresh every scheduled row's expected pot = monthly × active members
+    // Refresh every scheduled row's expected pot = monthly × active members.
+    // COUNT() is bigint; cast monthly_amount to numeric so "500.00" isn't
+    // coerced to bigint (which rejects the decimals).
     await exec(
       `UPDATE payout_schedule ps
-       SET expected_amount = $2 * (SELECT COUNT(*) FROM group_members
+       SET expected_amount = $2::numeric * (SELECT COUNT(*) FROM group_members
                                    WHERE group_id = $1 AND status = 'active')
        WHERE ps.group_id = $1 AND ps.cycle_number = $3 AND ps.status = 'scheduled'`,
       [groupId, g.monthly_amount, cycle]
